For #11365: Fixes UI tests for Settings
